[grid] rename GridPolar => GridAngle
