+++ packages/vx-grid/src/index.js
@@ -1,5 +1,5 @@
 export { default as GridRows } from './grids/GridRows';
 export { default as GridColumns } from './grids/GridColumns';
 export { default as Grid } from './grids/Grid';
-export { default as GridPolar } from './grids/GridPolar';
+export { default as GridAngle } from './grids/GridAngle';
 export { default as GridRadial } from './grids/GridRadial';
